requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.36.6
botocore==1.36.6
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.3.0
Django==5.1.5
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
load-dotenv==0.1.0
mysqlclient==2.2.7
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
vine==5.1.0
wcwidth==0.2.13
